TopModel Cleanup and Clark Transforms
Removes all the placeholders in the top level and creates a model library to contain electrical related blocks (ElectricalModelLib), starting with the Clark transform and the inverse Clark transform.

Takes also the opportunity to create in the test folder a test suite for the electrical model library and a (sanity) test case for the Clark transforms based on their assumed mathematical properties.